# docs/swagger.yaml
basePath: /
definitions:
...
  auth.InvitationPreview:
    properties:
      email:
        type: string
      grants_admin:
        description: |-
          GrantsAdmin reports that accepting will also activate a pending
          administrative grant, so the page can say what is being confirmed.
        type: boolean
      requires_password:
        type: boolean
    type: object
...
  handlers.AcceptInvitationRequest:
    properties:
      current_password:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - token
    type: object
...
  handlers.ChangeEmailRequest:
    properties:
      new_email:
        type: string
    required:
    - new_email
    type: object
...
host: localhost:8082
info:
  contact:
    name: EngineersMind
    url: https://github.com/engineersmind/emc-auth-server
  description: Standalone multi-tenant Identity Provider — email/password auth, JWT,
    refresh tokens, TOTP 2FA, SAML 2.0, RBAC, Admin UI, and AI/Agent security.
  license:
    name: MIT
  title: EMC Auth Server
  version: 1.5.0
paths:
...
  /api/v1/auth/accept-invitation:
    post:
      consumes:
      - application/json
      description: Sets the password for an invited account using the token from the
        invitation email, and marks the address verified. Single-use.
      parameters:
      - description: Invitation token and chosen password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AcceptInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accept an invitation
      tags:
      - AUTH
...
  /api/v1/auth/change-email:
    post:
      consumes:
      - application/json
      description: Sends a confirmation link to the new address. The account email
        changes only after the link is followed. Requires a valid access token.
      parameters:
      - description: New email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangeEmailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request an email change
      tags:
      - AUTH
  /api/v1/auth/confirm-email-change:
    get:
      description: Applies a pending email change using the token sent to the new
        address, and marks it verified. Single-use.
      parameters:
      - description: Email change token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm an email change
      tags:
      - AUTH
...
  /api/v1/auth/invitation:
    get:
      description: Returns what a live invitation token is for, WITHOUT consuming
        it, so the landing page can ask for a password only when there is one to set.
        requires_password is false for an account that already has credentials — an
        existing user being made an administrator, or an administrator being re-instated
        — and for those the link only confirms the grant. grants_admin reports that
        accepting will activate a pending administrative grant.
      parameters:
      - description: Invitation token from the email
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.InvitationPreview'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Inspect an invitation link
      tags:
      - AUTH
...
  /api/v1/auth/unblock-account:
    get:
      description: Lifts an automatic failed-attempt lockout using the token from
        the blocked-account email. Single-use. Admin blocks cannot be lifted this
        way.
      parameters:
      - description: Unblock token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unblock an account
      tags:
      - AUTH
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your_access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"